refactor: 컴포넌트 이름, 위치, 파일을 tsx 형식으로 변경
Unauthorized 컴포넌트 파일을 Error 파일 안에 함수형 컴포넌트로 옮김. MissingPage는 PageNotFound로 Error는 ServerError로 변경.
